Rework Serial CLI to require message termination
Messages are processed once CR or LF is detected or when they get
too long. This allows use of terminals that don't send the whole
message at once, such as PuTTY.

It also fixes some issues with timing - occasionally, messages got cut
with the previous solution, because the data transfer was too slow.